"GameDesc.js"_Implement mobile style
